+++ sac_verifier_logic.h
@@ -55,7 +55,7 @@ public:
   osuCrypto::PRNG prng_seed_global_;
 
   int time_eq_1;
-  int tot_matrix_multiplication_time;
+  long tot_matrix_multiplication_time;
 };
